requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-generativeai==0.8.3
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
lxml==4.9.3
gunicorn==21.2.0
